third_party/WebKit/LayoutTests/fast/forms/textarea/textarea-inputmode.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>
description('Tests the behavior of .inputMode of HTMLTextAreaElement.');

var textarea = document.createElement('textarea');

// .inputMode just reflect the corresponding attributes.
textarea.type = 'text';
shouldBe('textarea.inputMode', '""');
textarea.setAttribute('inputmode', '0');
shouldBe('textarea.inputMode', '"0"');
textarea.setAttribute('inputmode', 'abc');
shouldBe('textarea.inputMode', '"abc"');

textarea.inputMode = 'foo';
shouldBe('textarea.getAttribute("inputmode")', '"foo"');

textarea.inputMode = '';
shouldBe('textarea.getAttribute("inputmode")', '""');

// Null.
debug('Setting null to inputMode:');
textarea.inputMode = null;
shouldBe('textarea.inputMode', '"null"');
shouldBe('textarea.getAttribute("inputmode")', '"null"');
textarea.setAttribute('inputmode', null);
shouldBe('textarea.inputMode', '"null"');

// Undefined.
debug('Setting undefined to inputMode:');
textarea.inputMode = undefined;
shouldBe('textarea.inputMode', '"undefined"');
shouldBe('textarea.getAttribute("inputmode")', '"undefined"');
textarea.setAttribute('inputmode', undefined);
shouldBe('textarea.inputMode', '"undefined"');

// Non-string.
debug('Setting non-string to inputMode:');
textarea.inputMode = 256;
shouldBe('textarea.inputMode', '"256"');
shouldBe('textarea.getAttribute("inputmode")', '"256"');
textarea.setAttribute('inputmode', 256);
shouldBe('textarea.inputMode', '"256"');

</script>
</body>
</html>
