third_party/WebKit/LayoutTests/fast/forms/textarea/textarea-inputmode-expected.txt

Tests the behavior of .inputMode of HTMLTextAreaElement.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS textarea.inputMode is ""
PASS textarea.inputMode is "0"
PASS textarea.inputMode is "abc"
PASS textarea.getAttribute("inputmode") is "foo"
PASS textarea.getAttribute("inputmode") is ""
Setting null to inputMode:
PASS textarea.inputMode is "null"
PASS textarea.getAttribute("inputmode") is "null"
PASS textarea.inputMode is "null"
Setting undefined to inputMode:
PASS textarea.inputMode is "undefined"
PASS textarea.getAttribute("inputmode") is "undefined"
PASS textarea.inputMode is "undefined"
Setting non-string to inputMode:
PASS textarea.inputMode is "256"
PASS textarea.getAttribute("inputmode") is "256"
PASS textarea.inputMode is "256"
PASS successfullyParsed is true

TEST COMPLETE
